requirements.txt
pygame
ttkbootstrap
Pillow